refactor: Illegal 예외를 커스텀 예외로 일관성 확보
