+++ src/main/java/roomescape/domain/Reservation.java
@@ -6,6 +6,9 @@ import java.util.List;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.ToString;
+import roomescape.exception.DuplicateEntityException;
+import roomescape.exception.EntityNotFoundException;
+import roomescape.exception.RoomEscapeException;
 
 @Getter
 @EqualsAndHashCode(of = "id")
@@ -35,7 +38,7 @@ public class Reservation {
 
     private static void validatePastDateTime(LocalDate date, ReservationTime time) {
         if (time.isPast(date)) {
-            throw new IllegalArgumentException("이전 날짜로 예약 할 수 없습니다.");
+            throw new RoomEscapeException("이전 날짜로 예약 할 수 없습니다.");
         }
     }
 
@@ -46,13 +49,13 @@ public class Reservation {
 
     private static void validateTheme(Theme theme) {
         if (theme == null) {
-            throw new IllegalArgumentException("테마 정보는 비어있을 수 없습니다.");
+            throw new RoomEscapeException("테마 정보는 비어있을 수 없습니다.");
         }
     }
 
     private static void validateReservationDateTime(LocalDate date, ReservationTime time) {
         if (date == null || time == null) {
-            throw new IllegalArgumentException("예약 날짜 및 시간 정보는 비어있을 수 없습니다.");
+            throw new RoomEscapeException("예약 날짜 및 시간 정보는 비어있을 수 없습니다.");
         }
     }
 
@@ -84,7 +87,7 @@ public class Reservation {
 
     private void validateNotPast() {
         if (this.time.isPast(this.date)) {
-            throw new IllegalArgumentException("이미 지난 예약입니다.");
+            throw new RoomEscapeException("이미 지난 예약입니다.");
         }
     }
