+++ src/main/java/roomescape/domain/Theme.java
@@ -4,6 +4,7 @@ import java.net.URI;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.ToString;
+import roomescape.exception.RoomEscapeException;
 
 @Getter
 @EqualsAndHashCode(of = "id")
@@ -36,7 +37,7 @@ public class Theme {
 
     private static void validateThumbnailImageUrl(String thumbnailImageUrl) {
         if (isInvalidImageUrl(thumbnailImageUrl)) {
-            throw new IllegalArgumentException("올바른 이미지 주소 형식이 아닙니다.");
+            throw new RoomEscapeException("올바른 이미지 주소 형식이 아닙니다.");
         }
     }
 
@@ -45,32 +46,32 @@ public class Theme {
             URI uri = URI.create(thumbnailImageUrl);
             return uri.getScheme() == null || !uri.getScheme().startsWith("http");
         } catch (Exception e) {
-            throw new IllegalArgumentException("올바른 이미지 주소 형식이 아닙니다.");
+            throw new RoomEscapeException("올바른 이미지 주소 형식이 아닙니다.");
         }
     }
 
     private static void validateDescription(String description) {
         if (description == null || description.isBlank()) {
-            throw new IllegalArgumentException("설명은 필수 값입니다.");
+            throw new RoomEscapeException("설명은 필수 값입니다.");
         }
     }
 
     private static void validateName(String name) {
         if (name == null || name.isBlank()) {
-            throw new IllegalArgumentException("이름은 필수 값입니다.");
+            throw new RoomEscapeException("이름은 필수 값입니다.");
         }
     }
 
     public void deactivate() {
         if (!isActive) {
-            throw new IllegalArgumentException("이미 비활성화 된 테마입니다.");
+            throw new RoomEscapeException("이미 비활성화 된 테마입니다.");
         }
         this.isActive = false;
     }
 
     public void activate() {
         if (isActive) {
-            throw new IllegalArgumentException("이미 활성화 된 테마입니다.");
+            throw new RoomEscapeException("이미 활성화 된 테마입니다.");
         }
         this.isActive = true;
     }
