+++ src/main/java/roomescape/exception/ForbiddenException.java
@@ -1,4 +1,4 @@
-package roomescape.global.auth;
+package roomescape.exception;
 
 public class ForbiddenException extends RuntimeException {
     public ForbiddenException(String message) {
