+++ src/main/java/roomescape/exception/RoomEscapeException.java
@@ -1,0 +1,8 @@
+package roomescape.exception;
+
+public class RoomEscapeException extends RuntimeException {
+
+    public RoomEscapeException(String message) {
+        super(message);
+    }
+}
