+++ src/main/java/roomescape/global/web/RestExceptionHandler.java
@@ -13,16 +13,17 @@ import org.springframework.web.bind.MissingServletRequestParameterException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
-import roomescape.domain.DuplicateEntityException;
-import roomescape.domain.EntityNotFoundException;
-import roomescape.global.auth.ForbiddenException;
+import roomescape.exception.DuplicateEntityException;
+import roomescape.exception.EntityNotFoundException;
+import roomescape.exception.ForbiddenException;
+import roomescape.exception.RoomEscapeException;
 
 @RestControllerAdvice
 @Slf4j
 public class RestExceptionHandler {
 
-    @ExceptionHandler(value = {IllegalArgumentException.class, IllegalStateException.class})
-    public ResponseEntity<String> handleDuplicateEntityException(RuntimeException ex) {
+    @ExceptionHandler(RoomEscapeException.class)
+    public ResponseEntity<String> handleRoomEscapeException(RoomEscapeException ex) {
         return ResponseEntity.status(HttpStatus.BAD_REQUEST)
                 .body(ex.getMessage());
     }
