+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -4,8 +4,8 @@ import java.time.LocalTime;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.domain.DuplicateEntityException;
 import roomescape.domain.ReservationTime;
+import roomescape.exception.DuplicateEntityException;
 import roomescape.repository.ReservationTimeRepository;
 import roomescape.service.command.ReservationTimeCommand;
 import roomescape.service.result.ReservationTimeResult;
