+++ src/main/java/roomescape/service/ThemeService.java
@@ -3,8 +3,8 @@ package roomescape.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.domain.DuplicateEntityException;
 import roomescape.domain.Theme;
+import roomescape.exception.DuplicateEntityException;
 import roomescape.repository.ThemeRepository;
 import roomescape.service.command.ThemeRegisterCommand;
 import roomescape.service.result.ThemeRegisterResult;
