+++ src/test/java/integration/reservation/ReservationServiceIntTest.java
@@ -12,7 +12,7 @@ import java.util.concurrent.atomic.AtomicInteger;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import roomescape.domain.DuplicateEntityException;
+import roomescape.exception.DuplicateEntityException;
 import roomescape.service.ReservationService;
 import roomescape.service.command.ReservationCommand;
