+++ src/test/java/integration/theme/ThemeRepositoryTest.java
@@ -8,8 +8,8 @@ import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import roomescape.domain.DuplicateEntityException;
 import roomescape.domain.Theme;
+import roomescape.exception.DuplicateEntityException;
 import roomescape.repository.ThemeRepository;
 
 class ThemeRepositoryTest extends BaseIntegrationTest {
