+++ src/test/java/roomescape/service/ReservationTimeServiceTest.java
@@ -6,9 +6,9 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.time.LocalTime;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import roomescape.domain.DuplicateEntityException;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.TimeStatus;
+import roomescape.exception.DuplicateEntityException;
 import roomescape.service.command.ReservationTimeCommand;
 import roomescape.service.fake.FakeReservationTimeRepository;
 import roomescape.service.result.ReservationTimeResult;
